add png, split stories into files
