+++ frontend/src/metabase/static-viz/components/ComboChart/stories/Bar45DegreeLabels.stories.tsx
@@ -1,0 +1,13 @@
+import { data } from "../stories-data";
+
+import { Template, comboChartMeta, renderingContext } from "./shared";
+
+export default comboChartMeta;
+
+export const Bar45DegreeLabels = {
+  render: Template,
+  args: {
+    rawSeries: data.bar45DegreeLabels,
+    renderingContext,
+  },
+};
