+++ frontend/src/metabase/static-viz/components/ComboChart/stories/shared.tsx
@@ -1,0 +1,34 @@
+import type { StoryFn } from "@storybook/react";
+
+import {
+  measureTextHeight,
+  measureTextWidth,
+} from "metabase/static-viz/lib/text";
+import { color } from "metabase/ui/colors";
+import { DEFAULT_VISUALIZATION_THEME } from "metabase/visualizations/shared/utils/theme";
+import type { RenderingContext } from "metabase/visualizations/types";
+
+import type { StaticChartProps } from "../../StaticVisualization";
+import { StaticVisualization } from "../../StaticVisualization";
+
+export const comboChartMeta = {
+  title: "Viz/Static Viz/ComboChart",
+  component: StaticVisualization,
+};
+
+export const Template: StoryFn<StaticChartProps> = (args) => {
+  return (
+    <div style={{ border: "1px solid black", display: "inline-block" }}>
+      <StaticVisualization {...args} isStorybook />
+    </div>
+  );
+};
+
+export const renderingContext: RenderingContext = {
+  getColor: color,
+  measureText: (text, style) =>
+    measureTextWidth(text, Number(style.size), Number(style.weight)),
+  measureTextHeight: (_, style) => measureTextHeight(Number(style.size)),
+  fontFamily: "Lato",
+  theme: DEFAULT_VISUALIZATION_THEME,
+};
